update some  functions
对商品详情进行 显示；
将商品添加进购物车；
查看购物车；
更新问题总结.md

+++ src/main/java/com/ljy/mmall2/entity/Cart.java
@@ -32,7 +32,7 @@ public class Cart implements Serializable {
 
     private Integer quantity;
 
-    private Integer cost;
+    private Float cost;
 
     private Integer userId;
